# src/main/java/ar/edu/utn/frc/tup/lciii/model/CountryToSave.java
package ar.edu.utn.frc.tup.lciii.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CountryToSave {
    private int amountOfCountryToSave;
}